requirements.txt
absl-py==2.0.0
aiofiles==22.1.0
aiohttp==3.9.1
aiohttp-cors==0.7.0
aiorwlock==1.3.0
aiosignal==1.3.1
aiosqlite==0.19.0
alembic==1.13.1
altair==4.2.2
anyio==4.2.0
anyscale==0.5.131
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
attrs==23.1.0
avro==1.11.0
Babel==2.14.0
beautifulsoup4==4.12.2
bentoml==1.2.4
black==23.3.0
bleach==6.1.0
blessed==1.20.0
blinker==1.6.2
blis==0.7.11
boto3==1.34.11
botocore==1.34.46
build==1.0.3
CacheControl==0.14.0
cachetools==5.3.1
catalogue==2.0.10
cattrs==23.1.2
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.2.0
circus==0.18.0
cleanlab==2.3.1
cleo==2.1.0
click==8.1.7
click-option-group==0.5.6
cloudpathlib==0.16.0
cloudpickle==2.2.1
colorama==0.4.6
colorful==0.5.5
comm==0.1.4
confection==0.1.4
confluent-kafka==2.3.0
contextlib2==21.6.0
contourpy==1.2.0
coverage==7.4.0
crashtest==0.4.1
cryptography==41.0.7
cycler==0.12.1
cymem==2.0.8
databricks-cli==0.18.0
dataclasses==0.6
dataclasses-json==0.6.4
datasets==2.17.1
debugpy==1.8.0
decorator==5.1.1
deepmerge==1.1.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.8
distlib==0.3.8
distro==1.9.0
docker==6.1.3
dulwich==0.21.7
entrypoints==0.4
evaluate==0.4.1
executing==2.0.1
fastapi==0.95.2
fastavro==1.8.2
fastjsonschema==2.19.1
filelock==3.13.1
flake8==6.0.0
Flake8-pyproject==1.2.3
Flask==2.3.3
fonttools==4.47.0
fqdn==1.5.1
frozenlist==1.4.1
fs==2.4.16
fsspec==2023.10.0
furl==2.1.3
future==0.18.3
ghp-import==2.1.0
gitdb==4.0.10
GitPython==3.1.37
google-api-core==2.15.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
gotrue==2.4.2
gpustat==1.1.1
great-expectations==0.14.13
griffe==0.38.1
grpcio==1.60.0
gunicorn==20.1.0
h11==0.14.0
h5py==3.10.0
halo==0.0.31
hsfs==3.4.8
hsml==3.4.6
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.20.1
humanize==4.9.0
hyperopt==0.2.7
identify==2.5.33
idna==3.4
imageio==2.33.1
importlib-metadata==6.8.0
inflection==0.5.1
iniconfig==2.0.0
installer==0.7.0
ipykernel==6.26.0
ipython==8.17.2
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
isort==5.12.0
itsdangerous==2.1.2
jaraco.classes==3.3.1
javaobj-py3==0.4.3
jedi==0.19.1
Jinja2==3.0.3
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-events==0.9.0
jupyter-ydoc==0.2.5
jupyter_client==7.4.9
jupyter_core==5.5.0
jupyter_server==2.12.1
jupyter_server_fileid==0.9.1
jupyter_server_terminals==0.5.1
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.3
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keyring==24.3.1
kiwisolver==1.4.5
langchain==0.1.8
langchain-community==0.0.21
langchain-core==0.1.25
langchain-openai==0.0.6
langchain-pinecone==0.0.2
langcodes==3.3.0
langkit==0.0.30
langsmith==0.1.5
lazy_loader==0.3
lime==0.2.0.1
log-symbols==0.0.14
makefun==1.15.2
Mako==1.3.0
Markdown==3.3.7
markdown-it-py==3.0.0
MarkupSafe==2.0.1
marshmallow==3.20.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mergedeep==1.3.4
mistune==3.0.2
mkdocs==1.4.2
mkdocs-autorefs==0.5.0
mkdocstrings==0.21.2
mkdocstrings-python==1.7.5
mlflow==2.10.2
mock==5.1.0
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.16
munkres==1.1.4
murmurhash==1.0.10
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.9.0
nbconvert==7.13.1
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
nodeenv==1.8.0
notebook==6.5.6
notebook_shim==0.2.3
numpy==1.24.3
numpyencoder==0.3.0
nvidia-ml-py==11.525.150
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.12.0
opencensus==0.11.3
opencensus-context==0.1.3
opentelemetry-api==1.20.0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-aiohttp-client==0.41b0
opentelemetry-instrumentation-asgi==0.41b0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opentelemetry-util-http==0.41b0
orderedmultidict==1.0.1
overrides==7.4.0
packaging==23.2
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.12.1
pexpect==4.8.0
phonenumbers==8.13.30
Pillow==9.5.0
pinecone-client==3.0.3
pip-requirements-parser==32.0.1
pip-tools==7.4.0
pkginfo==1.10.0
platformdirs==3.11.0
pluggy==1.3.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
postgrest==0.16.4
pre-commit==3.2.2
preshed==3.0.9
presidio-analyzer==2.2.353
prometheus-client==0.19.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-spy==0.3.14
py4j==0.10.9.7
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodomex==3.20.0
pydantic==1.10.13
pydeck==0.8.1b0
pyflakes==3.0.1
Pygments==2.16.1
PyHopsHive==0.6.4.1.dev0
pyhumps==1.6.1
pyjks==20.0.0
PyJWT==2.8.0
pymdown-extensions==10.4
PyMuPDF==1.24.3
PyMuPDFb==1.24.3
PyMySQL==1.1.0
pyparsing==2.4.7
pyphen==0.14.0
pyproject_hooks==1.0.0
pytest==7.3.1
pytest-cov==4.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2023.3.post1
pyupgrade==3.3.2
PyYAML==6.0.1
pyyaml_env_tag==0.1
pyzmq==24.0.1
querystring-parser==1.2.4
rapidfuzz==3.7.0
ray==2.7.0
realtime==1.0.4
referencing==0.30.2
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.3
rpds-py==0.10.3
rsa==4.9
ruamel.yaml==0.17.17
s3transfer==0.10.0
safetensors==0.4.2
schema==0.7.5
scikit-image==0.22.0
scikit-learn==1.2.2
scipy==1.11.4
seaborn==0.12.2
Send2Trash==1.8.2
sentence-transformers==2.3.1
sentencepiece==0.2.0
shellingham==1.5.4
simple-di==0.1.5
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
snorkel==0.9.9
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spinners==0.0.24
SQLAlchemy==1.4.48
sqlparse==0.4.4
srsly==2.4.8
stack-data==0.6.3
starlette==0.27.0
storage3==0.7.4
streamlit==1.27.0
StrEnum==0.4.15
supabase==2.4.5
supafunc==0.4.5
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
terminado==0.18.0
textstat==0.7.3
thinc==8.2.3
threadpoolctl==3.2.0
thrift==0.20.0
tifffile==2023.12.9
tiktoken==0.6.0
tinycss2==1.2.1
tldextract==5.1.1
tokenize-rt==5.2.0
tokenizers==0.15.2
toml==0.10.2
tomlkit==0.12.4
toolz==0.12.0
torch==2.0.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.38.0
trove-classifiers==2024.3.25
twofish==0.3.0
typer==0.9.0
types-python-dateutil==2.8.19.14
types-requests==2.31.0.20240218
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2023.3
tzlocal==5.0.1
uri-template==1.3.0
urllib3==1.26.18
uvicorn==0.25.0
vaderSentiment==3.3.2
validators==0.22.0
virtualenv==20.25.1
wasabi==1.1.2
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.9
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
whylabs-client==0.5.10
whylogs==1.3.24
whylogs-sketching==3.4.1.dev3
widgetsnbextension==4.0.9
wordcloud==1.9.2
wrapt==1.16.0
xattr==1.1.0
xxhash==3.4.1
y-py==0.6.2
yarl==1.9.4
ypy-websocket==0.8.4
zipp==3.17.0
